Create README.md
This project presents the source codes of a new version of the multi-objective particle swarm optimizer named the Diversity-enhanced fuzzy multi-objective particle swarm optimization (f-MOPSO/Div) algorithm. This algorithm is an improved version of our recently proposed f-MOPSO algorithm that has presented a novel approach to handle multi-objectivity of the optimization problems. In the proposed algorithm, a new characteristic of the particles in the objective space, which we named the “extremity,” is also evaluated, along with the Pareto-dominance, to appoint proper guides for the particles in the search space. Three improvements are applied to the f-MOPSO to mitigate its shortcomings, generating f-MOPSO/Div: (1) selecting the global best solution based on the diversity of the extreme solutions, (2) impeding the particles to be trapped in the local optima using a mutation scheme based on the dynamic probability, and (3) removing the pre-optimization process. 

The source codes are specifically written to solve any multi/many-objective optimization problems having three or more objectives.

For further information about the functionality of this algorithm and how this algorithm handles the optimization problems, please refer to the following reference:

Rezaei, F., Safavi, H.R. f-MOPSO/Div: an improved extreme-point-based multi-objective PSO algorithm applied to a socio-economic-environmental conjunctive water use problem. Environ Monit Assess 192, 767 (2020). https://doi.org/10.1007/s10661-020-08727-y.

Please cite this article upon using this source code.